Portability changes 20250415

+++ dialog.hpp
@@ -51,8 +51,8 @@ struct DialogTemplate {
 	static inline void Alignment(BYTE* &p, SizeT &c, SizeT al) {
 		c += al; // 大きい分には問題ないのでサイズ計算用にはアライン分を足す
 		if (p) {
-			ULONGLONG n = al-1;
-			p = (BYTE*) ((intptr_t)(p + n) & ~n);
+			uintptr_t n = al-1;
+			p = (BYTE*) (((uintptr_t)p + n) & ~n);
 		}
 	}
